requirements.txt
asgiref==3.5.2
audioread==2.1.9
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tinytag==1.8.1
tzdata==2022.1
urllib3==1.26.9
